New metadynamics implementation without boundary conditions, grid based metadynamics in 1D and 2D, new shared bias implementation
