+++ adaptive_sampling/sampling_tools/metadynamics.py
@@ -1,11 +1,7 @@
-import os, time
 import numpy as np
-from typing import Tuple
-from .enhanced_sampling import EnhancedSampling
-from .utils import diff, sum
-from ..units import *
-from ..processing_tools.thermodynamic_integration import integrate
-
+from adaptive_sampling.sampling_tools.enhanced_sampling import EnhancedSampling
+from adaptive_sampling.units import *
+from .utils import correct_periodicity
 
 class WTM(EnhancedSampling):
     """Well-Tempered Metadynamics
@@ -18,208 +14,227 @@ class WTM(EnhancedSampling):
         hill_height: height of Gaussian hills in kJ/mol
         hill_std: standard deviation of Gaussian hills in units of the CVs (can be Bohr, Degree, or None)
         hill_drop_freq: frequency of hill creation in steps
-        well_tempered_temp: effective temperature for WTM, if None, hills are not scaled down (normal metadynamics)
+        well_tempered_temp: effective temperature for WTM, if np.inf, hills are not scaled down (normal metadynamics)
+        bias_factor: bias factor for WTM, if not None, overwrites `well_tempered_temp` 
         force_from_grid: forces are accumulated on grid for performance (recommended),
                          if False, forces are calculated from sum of Gaussians in every step
-        estimator: if "TI", PMF is estimated from integral of bias force, else PMF directly estimated from force
     """
 
     def __init__(
         self,
-        hill_height: float,
-        hill_std: list,
         *args,
-        hill_drop_freq: int = 20,
-        well_tempered_temp: float = 3000.0,
+        hill_height: float = -1,
+        hill_std: np.array = -1,
+        hill_drop_freq: int = 100,
+        well_tempered_temp: float = np.inf,
+        bias_factor: float = None,
         force_from_grid: bool = True,
-        estimator: str = "Potential",
+        verbose: bool = False,
         **kwargs,
     ):
         super().__init__(*args, **kwargs)
 
-        if int(hill_drop_freq) <= 0:
-            raise ValueError(" >>> Error: Update interval has to be int > 0!")
-
         if hill_height <= 0:
-            raise ValueError(" >>> Error: Gaussian height for MtD has to be > 0!")
+            raise ValueError(" >>> Error: Hill height for WTM has to be > 0!")
 
-        if well_tempered_temp is None:
-            if self.verbose:
-                print(
-                    " >>> Info: Well-tempered scaling of MtD hill_height switched off"
-                )
-        elif well_tempered_temp <= 0:
+        hill_std = np.array([hill_std]) if not hasattr(hill_std, "__len__") else hill_std # convert to array if float
+        if (hill_std <= 0).any():
+            raise ValueError(" >>> Error: Hill standard deviation for WTM has to be > 0!")        
+
+        if int(hill_drop_freq) <= 0:
+            raise ValueError(" >>> Error: Update interval for WTM has to be int > 0!")
+
+        if bias_factor is None and well_tempered_temp is None:
+            raise ValueError(
+                " >>> Error: Either bias_factor or well_tempered_temp has to be set!"
+            )
+
+        # general MtD parameters
+        self.beta = 1.0 / (kB_in_atomic * self.equil_temp)
+        self.verbose = verbose
+        self.force_from_grid = force_from_grid
+        self.hill_drop_freq = hill_drop_freq
+
+        # Well-Tempered parameters
+        if bias_factor is not None:
+            self.bias_factor = bias_factor
+            self.well_tempered_temp = self.bias_factor * self.equil_temp - self.equil_temp
+        else:
+            self.well_tempered_temp = well_tempered_temp
+            self.bias_factor = self.well_tempered_temp / self.equil_temp + 1.0 
+        self.well_tempered = False if self.well_tempered_temp == np.inf else True
+        self.wtm_prefac = (self.equil_temp + self.well_tempered_temp) / self.well_tempered_temp if self.well_tempered else 1.0
+        if not self.well_tempered and self.verbose:
+            print(
+                " >>> Info: Well-tempered scaling of MtD hill_height switched off"
+            )
+        elif self.well_tempered_temp <= 0:
             raise ValueError(
                 " >>> Error: Effective temperature for Well-Tempered MtD has to be > 0!"
             )
 
-        hill_std = [hill_std] if not hasattr(hill_std, "__len__") else hill_std
-
-        self.hill_height = hill_height / atomic_to_kJmol
+        # hill parameters
         self.hill_std = self.unit_conversion_cv(np.asarray(hill_std))[0]
-        self.hill_var = self.hill_std * self.hill_std
-        self.hill_drop_freq = int(hill_drop_freq)
-        self.well_tempered_temp = well_tempered_temp
-        self.well_tempered = False if well_tempered_temp == None else True
-        self.force_from_grid = force_from_grid
-        self.estimator = estimator
+        self.hill_height = hill_height / atomic_to_kJmol
+        self.hills_center = []
+        self.hills_height = []
+        self.hills_std = []
 
-        self.metapot = np.zeros_like(self.histogram)
-        self.center = []
-        self.bias_factor = (self.equil_temp + self.well_tempered_temp) / self.equil_temp if self.well_tempered_temp < np.inf else -1
-        self.beta = 1.0 / (kB_in_atomic * self.equil_temp)
-
+        # store results
+        self.metapot = np.copy(self.histogram)
+        self.bias_pot = 0.0
         self.mtd_rct = 0.0
-        self.rew_traj = []
+        self.bias_pot_traj = []
+        self.rct_traj = []
 
-    def step_bias(self, write_output: bool = True, write_traj: bool = True, **kwargs):
+        if self.verbose:
+            print(" >>> INFO: MtD Parameters:")
+            print("\t ---------------------------------------------")
+            print(f"\t Hill std:\t{self.hill_std}")
+            print(f"\t Hill height:\t{self.hill_height * atomic_to_kJmol} kJ/mol")
+            print(f"\t Bias factor:\t{self.bias_factor}\t\t({self.well_tempered_temp} K)")
+            print(f"\t Read force:\t{self.force_from_grid}")
+            print("\t ---------------------------------------------")
 
-        md_state = self.the_md.get_sampling_data()
-        (xi, delta_xi) = self.get_cv(**kwargs)
+    def step_bias(
+        self,
+        traj_file: str = "CV_traj.dat",
+        out_file: str = "mtd.out",
+        restart_file: str = "restart_mtd",
+        **kwargs,
+    ) -> np.array:
+        """Apply MtD bias to MD
 
-        bias_force = np.zeros_like(md_state.forces)
+        Returns:
+            bias_force: Bias force that has to be added to system forces
+        """
 
-        mtd_force = self.get_wtm_force(xi)
-        if self._check_boundaries(xi):
+        self.md_state = self.the_md.get_sampling_data()
+        (cv, grad_cv) = self.get_cv(**kwargs)
 
-            bink = self.get_index(xi)
-            self.histogram[bink[1], bink[0]] += 1
-
-            for i in range(self.ncoords):
-                bias_force += mtd_force[i] * delta_xi[i]
-            
-            self.rew_traj.append(np.exp(self.beta*(self.metapot[bink[1], bink[0]]-self.mtd_rct)))
-        else:
-            self.rew_traj.append(np.nan)
-
-        bias_force += self.harmonic_walls(xi, delta_xi)  # , 1.6 * self.hill_std)
-
-        self.traj = np.append(self.traj, [xi], axis=0)
-        self.temp.append(md_state.temp)
-        self.epot.append(md_state.epot)
+        # get mtd bias force
+        bias_force = self.harmonic_walls(cv, grad_cv)
+        mtd_forces = self.mtd_bias(cv)
+        for i in range(self.ncoords):
+            bias_force += mtd_forces[i] * grad_cv[i]
 
         # correction for kinetics
         if self.kinetics:
-            self._kinetics(delta_xi)
+            self._kinetics(grad_cv)
+
+        # store biased histogram along CV for output
+        if out_file and self._check_boundaries(cv):
+            bink = self.get_index(cv)
+            self.histogram[bink[1], bink[0]] += 1
 
         # shared-bias metadynamics
         if self.shared:
-            self.shared_bias(xi, **kwargs)
+            self.shared_bias(**kwargs)
 
-        if self.the_md.step % self.out_freq == 0:
-            # write output
+        # Save values for traj output
+        if traj_file:
+            self.traj = np.append(self.traj, [cv], axis=0)
+            self.epot.append(self.md_state.epot)
+            self.temp.append(self.md_state.temp)
+            self.bias_pot_traj.append(self.bias_pot)
+            self.rct_traj.append(self.mtd_rct)
 
-            if write_traj:
-                self.write_traj()
-
-            if write_output:
-                self.get_pmf()
+        # Write output
+        if self.md_state.step % self.out_freq == 0:
+            if traj_file and len(self.traj) >= self.out_freq:
+                self.write_traj(filename=traj_file)
+            if out_file:
+                self.pmf = self.get_pmf()
                 output = {
                     "hist": self.histogram,
-                    "free energy": self.pmf,
-                    "metapot": self.metapot * atomic_to_kJmol,
+                    "free energy": self.pmf * atomic_to_kJmol,
+                    "MtD Pot": self.metapot * atomic_to_kJmol,
                 }
-
-                self.write_output(output, filename="wtm.out")
-                self.write_restart()
+                self.write_output(output, filename=out_file)
+            if restart_file:
+                self.write_restart(filename=restart_file)
 
         return bias_force
 
-    def get_pmf(self):
-        if self.estimator == "TI" and self.ncoords == 1:
-            # PMF from integration of bias force
-            self.pmf[0, :], _ = integrate(
-                -self.bias[0][0],
-                self.dx,
-                equil_temp=self.equil_temp,
-                method="trapezoid",
-            )
+    def get_pmf(self) -> np.array:
+        """Calculate current PMF estimate on `self.grid`
 
-        else:
-            # PMF from bias potential
-            self.pmf = -self.metapot
-
+        Returns:
+            pmf: current PMF estimate from Mtd kernels
+        """
+        pmf = -self.metapot
         if self.well_tempered:
-            self.pmf *= (
-                self.equil_temp + self.well_tempered_temp
-            ) / self.well_tempered_temp
-        self.pmf *= atomic_to_kJmol
-        self.pmf -= self.pmf.min()
+            pmf *= self.wtm_prefac
+        pmf -= pmf.min()
+        return pmf
+
+    def write_traj(self, filename="CV_traj.dat"):
+        data = {
+            "Epot [Ha]": self.epot,
+            "T [K]": self.temp,
+            "Biaspot [Ha]": self.bias_pot_traj,
+            "C(t) [Ha]": self.rct_traj,
+        }
+        self._write_traj(data, filename=filename)
+
+        # Reset trajectories to save memory
+        self.traj = np.array([self.traj[-1]])
+        self.epot = [self.epot[-1]]
+        self.temp = [self.temp[-1]]
+        self.bias_pot_traj = [self.bias_pot_traj[-1]]
+        self.rct_traj = [self.rct_traj[-1]]
 
     def shared_bias(
         self,
-        xi,
         sync_interval: int = 50,
         mw_file: str = "shared_bias",
         n_trials: int = 10,
     ):
-        """TODO"""
+        """ Multiple walker shared bias implementation for metadynamics
+        """
+        import os, time
         md_state = self.the_md.get_sampling_data()
         if md_state.step == 0:
-            if self.verbose:
-                print(
-                    " >>> Info: Creating a new instance for shared-bias metadynamics."
-                )
-                print(
-                    " >>> Info: Data of local walker stored in `restart_wtm_local.npz`."
-                )
-
-            # create seperate restart file with local data only
-            self._write_restart(
-                filename="restart_wtm_local",
-                hist=self.histogram,
-                force=self.bias,
-                metapot=self.metapot,
-                centers=self.center,
-            )
 
             if sync_interval % self.hill_drop_freq != 0:
                 raise ValueError(
-                    " >>> Fatal Error: Sync interval for shared-bias WTM has to divisible through the frequency of hill creation!"
+                    " >>> ERROR: Sync interval for shared-bias WTM has to divisible through the frequency of hill creation!"
                 )
 
-            self.len_center_last_sync = len(self.center)
-            self.metapot_last_sync = np.copy(self.metapot)
-            self.bias_last_sync = np.copy(self.bias)
-            self.traj_last_sync = np.zeros(shape=(sync_interval, len(xi)))
+            if self.verbose:
+                print(" >>> Info: Creating a new instance for shared-bias metadynamics.")
+                print(" >>> Info: Data of local walker stored in `restart_wtm_local.npz`.")
+
+            # create seperate restart file with local data only
+            self.write_restart(
+                filename="restart_wtm_local",
+            )
+
+            self.num_hills_last_sync = len(self.hills_center)
+            self.hist_last_sync = np.copy(self.histogram)
             if not os.path.isfile(mw_file + ".npz"):
                 if self.verbose:
-                    print(
-                        f" >>> Info: Creating buffer file for shared-bias metadynamics: `{mw_file}.npz`."
-                    )
-                self._write_restart(
-                    filename=mw_file,
-                    hist=self.histogram,
-                    force=self.bias,
-                    metapot=self.metapot,
-                    centers=self.center,
-                )
+                    print(f" >>> Info: Creating buffer file for shared-bias metadynamics: `{mw_file}.npz`.")
+                self.write_restart(filename=mw_file)
                 os.chmod(mw_file + ".npz", 0o444)
             elif self.verbose:
-                print(
-                    f" >>> Info: Syncing with existing buffer file for shared-bias metadynamics: `{mw_file}.npz`."
-                )
+                print(f" >>> Info: Syncing with existing buffer file for shared-bias metadynamics: `{mw_file}.npz`.")
+        
+        if md_state.step % sync_interval == 0:
 
-        count = md_state.step % sync_interval
-        self.traj_last_sync[count] = xi
-        if count == sync_interval - 1:
+            # get new hills
+            new_center = np.asarray(self.hills_center)[self.num_hills_last_sync:]
+            new_height = np.asarray(self.hills_height)[self.num_hills_last_sync:]
+            new_std = np.asarray(self.hills_std)[self.num_hills_last_sync:]
 
-            new_center = self.center[self.len_center_last_sync :]
-            delta_hist = np.zeros_like(self.histogram)
-            for x in self.traj_last_sync:
-                if self._check_boundaries(x):
-                    bink = self.get_index(x)
-                    delta_hist[bink[1], bink[0]] += 1
-
-            delta_bias = self.bias - self.bias_last_sync
-            delta_metapot = self.metapot - self.metapot_last_sync
+            delta_hist = self.histogram - self.hist_last_sync
 
             self._update_wtm(
                 "restart_wtm_local",
                 delta_hist,
-                delta_bias,
-                delta_metapot,
                 new_center,
+                new_height,
+                new_std,
             )
 
             trial = 0
@@ -232,19 +247,18 @@ class WTM(EnhancedSampling):
                     self._update_wtm(
                         mw_file,
                         delta_hist,
-                        delta_bias,
-                        delta_metapot,
                         new_center,
+                        new_height,
+                        new_std,
                     )
                     self.restart(filename=mw_file)
                     os.chmod(mw_file + ".npz", 0o444)  # other walkers can access again
-
-                    # recalculates `self.metapot` and `self.bias` to ensure convergence of WTM potential
-                    self._update_metapot_from_centers()
-
-                    self.metapot_last_sync = np.copy(self.metapot)
-                    self.bias_last_sync = np.copy(self.bias)
-                    self.len_center_last_sync = len(self.center)
+                    self.num_hills_last_sync = len(self.hills_center)
+                    self.hist_last_sync = self.histogram
+                    if self.verbose:
+                        print(
+                            f" >>> Info: Synced bias with `{mw_file}.npz` after {trial} attempts."
+                        )
                     break
 
                 elif trial < n_trials:
@@ -257,235 +271,46 @@ class WTM(EnhancedSampling):
                     raise Exception(
                         f" >>> Fatal Error: Failed to sync bias with `{mw_file}.npz`."
                     )
-
-    def get_wtm_force(self, xi: np.ndarray) -> list:
-        """compute well-tempered metadynamics bias force from superposition of gaussian hills
-
-        Args:
-            xi: state of collective variable
-
-        Returns:
-            bias_force: bias force from metadynamics
-        """
-        if self.the_md.step % self.hill_drop_freq == 0:
-            self.center.append(xi[0]) if self.ncoords == 1 else self.center.append(xi)
-
-        is_in_bounds = (xi <= self.maxx).all() and (xi >= self.minx).all()
-        bias_force = [0, 0]
-        if is_in_bounds:
-            bias_force = self._accumulate_wtm_force(xi)
-            self.get_rct()
-
-        if not is_in_bounds or not self.force_from_grid:
-            bias_force, _ = self._analytic_wtm_force(xi)
-
-        return bias_force
-
-    def _accumulate_wtm_force(self, xi: np.ndarray) -> list:
-        """accumulate metadynamics potential and its gradient on grid
-
-        Args:
-            xi: state of collective variable
-
-        Returns:
-            bias_force: bias force from metadynamics
-        """
-        if self._check_boundaries(xi):
-            bink = self.get_index(xi)
-            if self.the_md.step % self.hill_drop_freq == 0:
-                if self.ncoords == 1:
-
-                    if self.well_tempered:
-                        w = self.hill_height * np.exp(
-                            -self.metapot[bink[1], bink[0]]
-                            / (kB_in_atomic * self.well_tempered_temp)
-                        )
-                    else:
-                        w = self.hill_height
-
-                    dx = diff(self.grid[0], xi[0], self.periodicity[0])
-                    epot = w * np.exp(-(dx * dx) / (2.0 * self.hill_var[0]))
-                    self.metapot[0] += epot
-                    self.bias[0][0] -= epot * dx / self.hill_var[0]
-
-                    self._smooth_boundary_grid(xi, w)
-
-                else:
-                    # TODO: implement for 2D
-                    raise NotImplementedError(
-                        " >>> Error: metadynamics currently only supported for 1D coordinates"
-                    )
-            return [self.bias[i][bink[1], bink[0]] for i in range(self.ncoords)]
-
-        else:
-            return [0.0 for _ in range(self.ncoords)]
-
-    def _analytic_wtm_force(self, xi: np.ndarray) -> Tuple[list, float]:
-        """compute analytic WTM bias force from sum of gaussians hills
-
-        Args:
-            xi: state of collective variable
-
-        Returns:
-            bias_force: bias force from metadynamics
-        """
-        local_pot = 0.0
-        bias_force = [0.0 for _ in range(self.ncoords)]
-
-        if len(self.center) == 0:
-            if self.verbose:
-                print(" >>> Warning: no metadynamics hills stored")
-            return bias_force, local_pot
-
-        if self.ncoords == 1:
-
-            dx = diff(xi[0], np.asarray(self.center), self.periodicity[0])
-            ind = np.ma.indices((len(self.center),))[0]
-            ind = np.ma.masked_array(ind)
-            ind[abs(dx) > 3 * self.hill_std[0]] = np.ma.masked
-
-            # can get slow in long run, so only iterate over significant elements
-            for i in np.nditer(ind.compressed(), flags=["zerosize_ok"]):
-
-                if self.well_tempered:
-                    w = self.hill_height * np.exp(
-                        -local_pot / (kB_in_atomic * self.well_tempered_temp)
-                    )
-                else:
-                    w = self.hill_height
-
-                epot = w * np.exp(-(dx[i] * dx[i]) / (2.0 * self.hill_var[0]))
-                local_pot += epot
-                bias_force[0] -= epot * dx[i] / self.hill_var[0]
-
-                local_pot, bias_force = self._smooth_boundary_analytic(
-                    xi, w, self.center[i], local_pot, bias_force
-                )
-        else:
-            # TODO: implement for 2D
-            pass
-
-        return bias_force, local_pot
-
-    def _smooth_boundary_grid(self, xi: np.ndarray, w: float):
-        """ensures linear profile of bias potential at boundary
-
-        see: Crespo et al. (2010), https://doi.org/10.1103/PhysRevE.81.055701
-
-        Args:
-            xi: Collective variable
-        """
-        chi1 = 3.0 * self.hill_std
-
-        if self.ncoords == 1:
-            dx_list = []
-            w_list = []
-
-            if (
-                diff(self.maxx[0], chi1[0], self.periodicity[0])
-                <= xi[0]
-                <= self.maxx[0]
-            ):
-                center_out = self.maxx[0] + diff(
-                    self.maxx[0], xi[0], self.periodicity[0]
-                )
-                dx_list.append(diff(self.grid[0], center_out, self.periodicity[0]))
-                w_list.append(w)
-            else:
-                # TODO: scale height of hills to ensure flat potential at boundary
-                pass
-
-            if self.minx[0] <= xi[0] <= sum(self.minx[0], chi1[0], self.periodicity[0]):
-                center_out = self.minx[0] - diff(
-                    xi[0], self.minx[0], self.periodicity[0]
-                )
-                dx_list.append(diff(self.grid[0], center_out, self.periodicity[0]))
-                w_list.append(w)
-            else:
-                # TODO: scale height of hills to ensure flat potential at boundary
-                pass
-
-            for dx, W in zip(dx_list, w_list):
-                epot = W * np.exp(-(dx * dx) / (2.0 * self.hill_var[0]))
-                self.metapot[0] += epot
-                self.bias[0][0] -= epot * dx / self.hill_var[0]
-
-        else:
-            # TODO: implement for 2D
-            pass
-
-    def _smooth_boundary_analytic(self, xi, w, center, local_pot, bias_force):
-        """ensures linear profile at boundary
-
-        see: Crespo et al. (2010), https://doi.org/10.1103/PhysRevE.81.055701
-
-        Args:
-            xi: Collective variable
-        """
-        chi1 = 3.0 * self.hill_std
-
-        if self.ncoords == 1:
-            if (
-                diff(self.maxx[0], chi1[0], self.periodicity[0])
-                <= center
-                <= self.maxx[0]
-            ):
-                center_out = self.maxx[0] + diff(
-                    self.maxx[0], center, self.periodicity[0]
-                )
-                val = diff(xi[0], center_out, self.periodicity[0])
-                epot = w * np.exp(-(val * val) / (2.0 * self.hill_var[0]))
-                local_pot += epot
-                bias_force[0] -= epot * val / self.hill_var[0]
-            else:
-                # TODO: scale height of hills to ensure flat potential at boundary
-                pass
-
-            if (
-                self.minx[0]
-                <= center
-                <= sum(self.minx[0], chi1[0], self.periodicity[0])
-            ):
-                center_out = self.minx[0] - diff(
-                    center, self.minx[0], self.periodicity[0]
-                )
-                val = diff(xi[0], center_out, self.periodicity[0])
-                epot = w * np.exp(-(val * val) / (2.0 * self.hill_var[0]))
-                local_pot += epot
-                bias_force[0] -= epot * val / self.hill_var[0]
-            else:
-                # TODO: scale height of hills to ensure flat potential at boundary
-                pass
-
-        return local_pot, bias_force
-
-    def _update_wtm(self, filename, hist, bias, metapot, center):
+            if self.ncoords == 1:
+                grid_full = np.asarray(self.grid[0]).reshape((-1,1))
+            elif self.ncoords == 2:
+                xx, yy = np.meshgrid(self.grid[0], self.grid[1])
+                grid_full = np.stack([xx.flatten(), yy.flatten()], axis=1)
+            
+            # recalculates `self.metapot` and `self.bias` to ensure convergence of WTM potential
+            shape = self.metapot.shape
+            self.metapot = np.zeros_like(self.metapot).flatten()
+            self.bias = np.zeros_like(self.bias).reshape((len(self.metapot), self.ncoords))
+            for i, bin_coords in enumerate(grid_full): 
+                pot, der = self.calc_gaussians(bin_coords, requires_grad=True)
+                self.metapot[i] = np.sum(pot)
+                self.bias[i] = der
+            self.metapot = self.metapot.reshape(shape)
+            self.bias = self.bias.reshape((self.ncoords,) + shape)
+                
+    def _update_wtm(self, 
+                    filename: str, 
+                    delta_hist: np.array, 
+                    new_hill_centers: np.array, 
+                    new_hill_heights: np.array, 
+                    new_hill_stds: np.array):
+        """updates shared bias buffer"""
         with np.load(f"{filename}.npz") as data:
-            new_hist = data["hist"] + hist
-            new_bias = data["force"] + bias
-            new_metapot = data["metapot"] + metapot
-            new_centers = np.append(data["centers"], center)
+            new_hist = data["hist"] + delta_hist
+            new_centers = np.append(data["center"], new_hill_centers)
+            new_heights = np.append(data["height"], new_hill_heights)
+            new_stds = np.append(data["sigma"], new_hill_stds)
 
         self._write_restart(
             filename=filename,
             hist=new_hist,
-            force=new_bias,
-            metapot=new_metapot,
-            centers=new_centers,
+            height=new_heights,
+            center=new_centers,
+            sigma=new_stds,
         )
 
-    def get_rct(self):
-        """get reweighting factor for metadynamics
-        """
-        minusBetaF = self.beta * self.bias_factor / (self.bias_factor - 1.) if self.bias_factor != -1 else self.beta
-        minusBetaFplusV = self.beta / (self.bias_factor - 1.) if self.bias_factor != -1 else 0.0
-        max_bias = self.metapot.max() # to avoid overflow
-        Z_0 = np.sum(np.exp(minusBetaF * (self.metapot - max_bias)))
-        Z_V = np.sum(np.exp(minusBetaFplusV * (self.metapot - max_bias))) 
-        self.mtd_rct = (1. / self.beta) * np.log(Z_0/Z_V) + max_bias
-
-    def write_restart(self, filename: str = "restart_wtm"):
-        """write restart file
+    def write_restart(self, filename: str = "restart_mtd"):
+        """Dumps state of MtD to restart file
 
         Args:
             filename: name of restart file
@@ -493,71 +318,158 @@ class WTM(EnhancedSampling):
         self._write_restart(
             filename=filename,
             hist=self.histogram,
-            force=self.bias,
-            metapot=self.metapot,
-            centers=self.center,
+            height=self.hills_height,
+            center=self.hills_center,
+            sigma=self.hills_std,
         )
 
-    def restart(self, filename: str = "restart_wtm"):
-        """restart from restart file
+    def restart(self, filename: str = "restart_mtd"):
+        """Restart MtD from previous simulation
 
         Args:
-            filename: name of restart file
+            filename: name of restart
         """
         try:
             data = np.load(filename + ".npz", allow_pickle=True)
         except:
-            raise OSError(f" >>> fatal error: restart file `{filename}.npz` not found!")
+            raise OSError(f" >>> ERROR: restart file {filename}.npz not found!")
 
+        # Load dictionary entries from restart file
         self.histogram = data["hist"]
-        self.bias = data["force"]
-        self.metapot = data["metapot"]
-        self.center = data["centers"].tolist()
+        self.hills_height = data["height"].tolist()
+        self.hills_center = data["center"].tolist()
+        self.hills_std = data["sigma"].tolist()
+        if not hasattr(self.hills_center[0], "__len__"):
+            self.hills_center = [np.array([c]) for c in self.hills_center]
+            self.hills_std = [np.array([std]) for std in self.hills_std]
+        if self.verbose and self.md_state.step % self.hill_drop_freq == 0 and not self.shared_bias:
+            print(f" >>> Info: Adaptive sampling restarted from {filename}!")
 
-        if self.verbose:
-            print(f" >>> Info: Adaptive sampling restartet from `{filename}.npz`!")
+    def mtd_bias(self, cv: np.array) -> np.array:
+        """Calculate MtD bias force from KDE of hills
+        which is updated according to `self.update_freq`
 
-    def _update_metapot_from_centers(self):
-        """recalculate metadynamics potential and bias force from stored centers"""
-        if self.ncoords > 1:
-            # TODO: implement for 2D
-            raise NotImplementedError(
-                " >>> Error: metadynamics currently only supported for 1D coordinates"
+        Args:
+            cv: new value of CV
+
+        Returns:
+            bias force: len(ncoords) array of bias forces
+        """
+        # get bias potential and forces
+        if self.force_from_grid and self._check_boundaries(cv):
+            idx = self.get_index(cv)
+            self.bias_pot = self.metapot[idx[1], idx[0]]
+            mtd_force = [self.bias[i][idx[1], idx[0]] for i in range(self.ncoords)]
+        else:
+            gaussians, derivative = self.calc_gaussians(cv, requires_grad=True)
+            self.bias_pot = np.sum(gaussians)
+            mtd_force = derivative
+
+        # MtD KDE update
+        if self.md_state.step % self.hill_drop_freq == 0:
+            self.update_kde(cv)
+
+        return mtd_force
+
+    def calc_gaussians(self, cv, requires_grad: bool = False) -> np.array:
+        """Get value of sum of gaussian hills
+
+        Args:
+            cv: value of CV where the kernels should be evaluated
+            requires_grad: if True, gradient of mtd potential is returned as second argument
+
+        Returns:
+            gaussians: values of gaussians at CV
+            kde_derivative: derivative of KDE, only if requires_grad
+        """
+
+        if len(self.hills_center) == 0:
+            if requires_grad:
+                return 0.0, np.zeros(self.ncoords)
+            return 0.0
+
+        # distance to kernel centers
+        s_diff = cv - np.asarray(self.hills_center)
+        for i in range(self.ncoords):
+            s_diff[:, i] = correct_periodicity(s_diff[:, i], self.periodicity[i])
+
+        # evaluate values of kernels at cv
+        gaussians = np.asarray(self.hills_height) * np.exp(
+            -0.5
+            * np.sum(np.square(np.divide(s_diff, np.asarray(self.hills_std))), axis=1)
+        )
+        if requires_grad:
+            derivative = np.sum(
+                -gaussians * np.divide(s_diff, np.square(np.asarray(self.hills_std))).T,
+                axis=1,
             )
+            return gaussians, derivative
 
-        self.metapot = np.zeros_like(self.metapot)
-        self.bias = np.zeros_like(self.bias)
+        return gaussians
 
-        for _, xi in enumerate(self.center):
-            xi = [xi]
-            if self._check_boundaries(xi):
-                bink = self.get_index(xi)
-                if self.well_tempered:
-                    w = self.hill_height * np.exp(
-                        -self.metapot[bink[1], bink[0]]
-                        / (kB_in_atomic * self.well_tempered_temp)
-                    )
-                else:
-                    w = self.hill_height
+    def update_kde(self, cv: np.array):
+        """on-the-fly update of kernel density estimation of pmf along CVs
 
-                dx = diff(self.grid[0], xi[0], self.periodicity[0])
-                epot = w * np.exp(-(dx * dx) / (2.0 * self.hill_var[0]))
-                self.metapot[0] += epot
-                self.bias[0][0] -= epot * dx / self.hill_var[0]
+        Args:
+            CV: new value of CVS
+        """
+        self._add_kernel(self.hill_height, cv, self.hill_std)
+        self._update_grid_potential(self.hills_height[-1], self.hills_center[-1], self.hills_std[-1])
+        self._get_rct()
 
-                self._smooth_boundary_grid(xi, w)
+    def _add_kernel(self, h_new: float, s_new: np.array, std_new: np.array):
+        """Add new Kernel to KDE
 
-    def write_traj(self):
-        """save trajectory for post-processing"""
-        data = {
-            "Epot [H]": self.epot,
-            "T [K]": self.temp,
-            "C(T)": self.rew_traj,
-        }
-        self._write_traj(data)
+        Args:
+            h_new: hills height
+            s_new: hills position
+            std_new: hills standard deviation
+        """
+        if self.well_tempered:
+            w = h_new * np.exp(
+                -self.bias_pot / (kB_in_atomic * self.well_tempered_temp)
+            )
+        else:
+            w = h_new
 
-        # reset trajectories to save memory
-        self.traj = np.array([self.traj[-1]])
-        self.epot = [self.epot[-1]]
-        self.temp = [self.temp[-1]]
-        self.rew_traj = [self.rew_traj[-1]]
+        self.hills_height.append(w)
+        self.hills_center.append(s_new)
+        self.hills_std.append(std_new)
+    
+    def _get_rct(self):
+        """get reweighting factor for metadynamics
+        """
+        minusBetaF = self.beta * self.bias_factor / (self.bias_factor - 1.) if self.well_tempered else self.beta
+        minusBetaFplusV = self.beta / (self.bias_factor - 1.) if self.well_tempered else 0.0
+        max_bias = self.metapot.max() # to avoid overflow
+        Z_0 = np.sum(np.exp(minusBetaF * (self.metapot - max_bias)))
+        Z_V = np.sum(np.exp(minusBetaFplusV * (self.metapot - max_bias))) 
+        self.mtd_rct = (1. / self.beta) * np.log(Z_0/Z_V) + max_bias
+
+    def _update_grid_potential(self, height_new: np.array, center_new: np.array, std_new: np.array):
+        """Add new kernel to `self.metapot` and `self.bias` 
+        
+        Args:
+            height_new: height of new kernel
+            center_new: center of new kernel
+            std_new: standard deviation of new kernel
+        """
+        if self.ncoords == 1:
+            grid_full = np.asarray(self.grid[0]).reshape((-1,1))
+        elif self.ncoords == 2:
+            xx, yy = np.meshgrid(self.grid[0], self.grid[1])
+            grid_full = np.stack([xx, yy], axis=-1)
+        else:
+            raise NotImplementedError(" >>> ERROR: MtD grid potential not implemented for ncoords > 2")
+
+        s_diff = np.rollaxis(grid_full - np.asarray(center_new), -1)
+        for i in range(self.ncoords):
+            s_diff[i,:] = correct_periodicity(s_diff[i,:], self.periodicity[i])
+        pot_new = np.asarray(height_new) * np.exp(
+            -0.5 * np.sum(np.square(np.divide(s_diff.T, np.asarray(std_new))), axis=-1)
+        ).T
+        self.metapot += pot_new
+        if self.force_from_grid:
+            for i in range(self.ncoords):
+                self.bias[i] -= pot_new * np.divide(s_diff[i], np.square(std_new[i]))
+
